CalenderHandlerModule/services/orchestrator/requirements.txt
fastapi
uvicorn
httpx
python-dateutil
google-auth
google-auth-oauthlib
google-api-python-client
